Rework of valued enums to c# style

+++ src/main/kotlin/ru/tesserakt/diskordin/core/data/json/request/PermissionsEditRequest.kt
@@ -2,7 +2,7 @@ package ru.tesserakt.diskordin.core.data.json.request
 
 
 data class PermissionsEditRequest(
-    val allow: Int,
-    val deny: Int,
+    val allow: Long,
+    val deny: Long,
     val type: String
 ) : JsonRequest()
